Move styles and tests
