+++ src/js/disability-benefits/526EZ/form-entry.jsx
@@ -1,5 +1,5 @@
 import '../../../platform/polyfills';
-import '../../../sass/disability-benefits.scss';
+import './sass/disability-benefits.scss';
 
 import startApp from '../../../platform/startup';
