+++ src/js/disability-benefits/526EZ/sass/disability-benefits.scss
@@ -4,10 +4,10 @@
 @import "~@department-of-veterans-affairs/jean-pants/sass/modules/m-process-list";
 @import "~@department-of-veterans-affairs/jean-pants/sass/modules/m-form-process";
 @import "~@department-of-veterans-affairs/jean-pants/sass/modules/m-progress-bar";
-@import "../platform/forms/sass/m-schemaform";
+@import "../../../../platform/forms/sass/m-schemaform";
 // @import "~@department-of-veterans-affairs/jean-pants/sass/modules/m-modal";
 // @import "~@department-of-veterans-affairs/jean-pants/sass/modules/m-omb-info";
-@import "../platform/forms/sass/m-form-confirmation";
+@import "../../../../platform/forms/sass/m-form-confirmation";
 
 
 .dashed-bullet {
