+++ src/js/disability-benefits/526EZ/tests/config/documentUpload.unit.spec.jsx
@@ -4,8 +4,8 @@ import sinon from 'sinon';
 import { mount } from 'enzyme';
 
 import { DefinitionTester, // selectCheckbox 
-} from '../../../../src/platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/disability-benefits/526EZ/config/form.js';
+} from '../../../../../platform/testing/unit/schemaform-utils.jsx';
+import formConfig from '../../config/form.js';
 import initialData from '../schema/initialData.js';
 
 const invalidDocumentData = {
