+++ src/js/disability-benefits/526EZ/tests/config/ratedDisabilities.unit.spec.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 
-import { DefinitionTester } from '../../../../src/platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../../src/js/disability-benefits/526EZ/config/form.js';
+import { DefinitionTester } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
+import formConfig from '../../config/form.js';
 import initialData from '../schema/initialData.js';
 
 describe('Disability benefits 526EZ -- Rated disabilities selection', () => {
